+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file.
 ### Added
  - title property to the Window element
  - color property to the Window element
+ - maximum/minimum properties to the SpinBox
 
 
 ## [0.0.4] - 2020-12-04